--- a/docker-compose.optimized.yml
+++ b/docker-compose.optimized.yml
@@ -95,6 +95,7 @@ services:
       - "8000:8000"
     environment:
       - CUDA_VISIBLE_DEVICES=0
+      - NVIDIA_VISIBLE_DEVICES=0
       - VLLM_MODEL=${VLLM_MODEL:-microsoft/DialoGPT-medium}
       - VLLM_HOST=0.0.0.0
       - VLLM_PORT=8000
@@ -163,7 +164,8 @@ services:
     ports:
       - "8001:8001"
     environment:
-      - CUDA_VISIBLE_DEVICES=0
+      - CUDA_VISIBLE_DEVICES=1
+      - NVIDIA_VISIBLE_DEVICES=1
       - QDRANT_HOST=qdrant
       - QDRANT_PORT=6333
       - POSTGRES_HOST=postgres