--- a/services/multimodal-worker/Dockerfile.optimized
+++ b/services/multimodal-worker/Dockerfile.optimized
@@ -8,6 +8,9 @@ FROM multimodal-base:latest as builder
 ENV PIP_NO_CACHE_DIR=1 \
     PIP_DISABLE_PIP_VERSION_CHECK=1
 
+# Switch to root for package installation
+USER root
+
 # Copy requirements first for better layer caching
 COPY requirements.txt .
 
@@ -18,14 +21,21 @@ RUN pip install --no-cache-dir --upgrade pip && \
 # Production stage
 FROM multimodal-base:latest as production
 
+# Switch to root temporarily for setup
+USER root
+
 # Copy Python dependencies from builder
-COPY --from=builder /home/appuser/.local /home/appuser/.local
+COPY --from=builder /usr/local/lib/python3.11/site-packages /usr/local/lib/python3.11/site-packages
+COPY --from=builder /usr/local/bin /usr/local/bin
 
 # Copy application code
 COPY --chown=appuser:appuser . .
 
 # Create cache directory
 RUN mkdir -p /app/cache && chown appuser:appuser /app/cache
+
+# Switch back to appuser
+USER appuser
 
 # Expose port
 EXPOSE 8001