--- a/services/retrieval-proxy/Dockerfile.optimized
+++ b/services/retrieval-proxy/Dockerfile.optimized
@@ -1,0 +1,40 @@
+# Optimized Retrieval Proxy Service Dockerfile
+# Uses shared base image and advanced caching
+
+# Build stage - use shared base image
+FROM multimodal-base:latest as builder
+
+# Set environment variables
+ENV PIP_NO_CACHE_DIR=1 \
+    PIP_DISABLE_PIP_VERSION_CHECK=1
+
+# Copy requirements first for better layer caching
+COPY requirements.txt .
+
+# Install Python dependencies in virtual environment for better isolation
+RUN python -m venv /opt/venv
+ENV PATH="/opt/venv/bin:$PATH"
+RUN pip install --no-cache-dir --upgrade pip && \
+    pip install --no-cache-dir -r requirements.txt
+
+# Production stage
+FROM multimodal-base:latest as production
+
+# Set environment variables
+ENV PATH="/opt/venv/bin:$PATH"
+
+# Copy virtual environment from builder stage
+COPY --from=builder /opt/venv /opt/venv
+
+# Copy application code
+COPY --chown=appuser:appuser . .
+
+# Expose port
+EXPOSE 8002
+
+# Health check
+HEALTHCHECK --interval=30s --timeout=10s --start-period=30s --retries=3 \
+    CMD curl -f http://localhost:8002/health || exit 1
+
+# Run the application
+CMD ["python", "main.py"]