--- a/services/user-management/Dockerfile.optimized
+++ b/services/user-management/Dockerfile.optimized
@@ -1,0 +1,31 @@
+# Optimized Dockerfile for User Management Service
+# Inherit from the shared base image
+FROM multimodal-base:latest
+
+# Switch back to root temporarily to create /app directory for appuser
+USER root
+RUN mkdir -p /app && chown -R appuser:appuser /app
+USER appuser
+
+# Set working directory
+WORKDIR /app
+
+# Copy requirements first for better caching
+COPY --chown=appuser:appuser requirements.txt .
+
+# Install Python dependencies in the virtual environment
+# This layer will be cached as long as requirements.txt doesn't change
+RUN pip install --no-cache-dir -r requirements.txt
+
+# Copy application code
+COPY --chown=appuser:appuser . .
+
+# Expose port
+EXPOSE 8006
+
+# Health check
+HEALTHCHECK --interval=30s --timeout=10s --start-period=60s --retries=3 \
+    CMD curl -f http://localhost:8006/health || exit 1
+
+# Run the application
+CMD ["python", "main.py"]