… docker-compose.ai-agents.yml
… docker-compose.functional.yml
… docs/README.md
… docs/docker-compose-cleanup-summary.md
… docs/docker-compose-usage-guide.md
… docs/gpu-configuration-guide.md